+++ source_gen/source_gen/source_gen/lib/src/generated_output.dart
@@ -17,7 +17,10 @@ class GeneratedOutput {
   GeneratedOutput(this.generator, this.output)
       : error = null,
         stackTrace = null,
-        assert(output != null && output.isNotEmpty);
+        assert(output != null),
+        assert(output.isNotEmpty),
+        // assuming length check is cheaper than simple string equality
+        assert(output.length == output.trim().length);
 
   GeneratedOutput.fromError(this.generator, this.error, this.stackTrace)
       : this.output = _outputFromError(error);